+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/CallPage/CallPage.tsx
@@ -1,5 +1,4 @@
 import Box from '@mui/material/Box';
-import {ErrorPanel} from '@wandb/weave/components/ErrorPanel';
 import {Loading} from '@wandb/weave/components/Loading';
 import {useViewTraceEvent} from '@wandb/weave/integrations/analytics/useViewEvents';
 import React, {FC, useCallback} from 'react';
@@ -8,11 +7,7 @@ import {useHistory} from 'react-router-dom';
 import {Button} from '../../../../../Button';
 import {Tailwind} from '../../../../../Tailwind';
 import {Browse2OpDefCode} from '../../../Browse2/Browse2OpDefCode';
-import {
-  TRACETREE_PARAM,
-  useClosePeek,
-  useWeaveflowCurrentRouteContext,
-} from '../../context';
+import {TRACETREE_PARAM, useWeaveflowCurrentRouteContext} from '../../context';
 import {FeedbackGrid} from '../../feedback/FeedbackGrid';
 import {makeRefCall} from '../../refs';
 import {isEvaluateOp} from '../common/heuristics';
@@ -26,6 +21,7 @@ import {CallDetails} from './CallDetails';
 import {CallOverview} from './CallOverview';
 import {CallSummary} from './CallSummary';
 import {CallTraceView, useCallFlattenedTraceTree} from './CallTraceView';
+import {NotFoundPanel} from '../../NotFoundPanel';
 
 export const CallPage: FC<{
   entity: string;
@@ -34,7 +30,6 @@ export const CallPage: FC<{
   path?: string;
 }> = props => {
   const {useCall} = useWFHooks();
-  const close = useClosePeek();
 
   const call = useCall({
     entity: props.entity,
@@ -45,16 +40,7 @@ export const CallPage: FC<{
   if (call.loading) {
     return <CenteredAnimatedLoader />;
   } else if (call.result === null) {
-    return (
-      <div style={{display: 'flex', flexDirection: 'column', height: '100%'}}>
-        <div style={{alignSelf: 'flex-end', margin: 10}}>
-          <Button icon="close" variant="ghost" onClick={close} />
-        </div>
-        <div style={{flex: 1}}>
-          <ErrorPanel title="Call not found" subtitle="" subtitle2="" />
-        </div>
-      </div>
-    );
+    return <NotFoundPanel title="Call not found" />;
   }
   return <CallPageInnerVertical {...props} call={call.result} />;
 };
